+++ src/components/App/Register/register.js
@@ -28,7 +28,7 @@ import VisibilityOff from '@material-ui/icons/VisibilityOff'
 
 import img from '../../../images/auth.svg'
 import { setup } from '../../../services/Auth'
-import oStyle from '../../../components/App/Register/registerStyle'
+import oStyle from '../../../services/css/registerStyle'
 import { registerUser, cardCheck } from '../../../store/actions'
 import GradientBtn from '../../UI/buttons/GradientBtn'
 import { checkPseudo, checkEmail, checkPassword } from '../../../utils'
